Fix typescript weirdness
