+++ tsconfig.json
@@ -2,7 +2,6 @@
   "extends": "@adonisjs/tsconfig/tsconfig.package.json",
   "compilerOptions": {
     "rootDir": "./",
-    "outDir": "./build",
-    "module": "Preserve"
+    "outDir": "./build"
   }
 }
